+++ src/util/linux_util.h
@@ -32,7 +32,7 @@ GPtrArray* rpt_lsof_collect(               const char * fqfn);
 GPtrArray* diagnose_open_failure_collect(  const char * fqfn, const char * msg, GPtrArray* collector);
 void       diagnose_open_failure_to_syslog(const char * fqfn, const char * msg);
 
-void       install_segv_handler(void);
+void       install_fatal_signal_handlers(void);
 
 void       init_accumulated_sleep();
 bool       recently_resumed_from_sleep_by_clocktime0(bool no_mutate, bool * detected_now_loc);